q:ddressing issues from qlty 01
