+++ .github/workflows/changesets.yml
@@ -1,3 +1,4 @@
+---
 name: Changesets
 
 on:
@@ -7,8 +8,8 @@ on:
 
 # These permissions are needed to interact with GitHub's API
 permissions:
-  contents: write # to create release (changesets/action)
-  pull-requests: write # to create pull request (changesets/action)
+  contents: write
+  pull-requests: write
 
 jobs:
   release:
