+++ .github/workflows/e2e-tests.yml
@@ -1,3 +1,4 @@
+---
 name: Playwright
 
 on:
