+++ .github/workflows/setup.yml
@@ -1,3 +1,4 @@
+---
 name: Setup
 on:
   workflow_call:
