+++ .github/workflows/vitest.yml
@@ -1,3 +1,4 @@
+---
 name: Vitest
 
 on:
